a little refactoring
