+++ src/main/java/com/badbones69/crazycrates/api/enums/Messages.java
@@ -6,13 +6,12 @@ import com.ryderbelserion.vital.core.util.StringUtil;
 import com.ryderbelserion.vital.paper.enums.Support;
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.command.CommandSender;
-import org.jetbrains.annotations.Nullable;
 import com.badbones69.crazycrates.config.ConfigManager;
-import com.badbones69.crazycrates.config.impl.messages.CommandKeys;
-import com.badbones69.crazycrates.config.impl.messages.CrateKeys;
-import com.badbones69.crazycrates.config.impl.messages.ErrorKeys;
-import com.badbones69.crazycrates.config.impl.messages.MiscKeys;
-import com.badbones69.crazycrates.config.impl.messages.PlayerKeys;
+import com.badbones69.crazycrates.config.impl.locale.CommandKeys;
+import com.badbones69.crazycrates.config.impl.locale.CrateKeys;
+import com.badbones69.crazycrates.config.impl.locale.ErrorKeys;
+import com.badbones69.crazycrates.config.impl.locale.MiscKeys;
+import com.badbones69.crazycrates.config.impl.locale.PlayerKeys;
 import me.clip.placeholderapi.PlaceholderAPI;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
