+++ src/main/java/com/badbones69/crazycrates/api/enums/Properties.java
@@ -4,11 +4,11 @@ import ch.jalu.configme.configurationdata.ConfigurationData;
 import ch.jalu.configme.properties.Property;
 import ch.jalu.configme.resource.PropertyReader;
 import com.badbones69.crazycrates.config.impl.ConfigKeys;
-import com.badbones69.crazycrates.config.impl.messages.CommandKeys;
-import com.badbones69.crazycrates.config.impl.messages.CrateKeys;
-import com.badbones69.crazycrates.config.impl.messages.ErrorKeys;
-import com.badbones69.crazycrates.config.impl.messages.MiscKeys;
-import com.badbones69.crazycrates.config.impl.messages.PlayerKeys;
+import com.badbones69.crazycrates.config.impl.locale.CommandKeys;
+import com.badbones69.crazycrates.config.impl.locale.CrateKeys;
+import com.badbones69.crazycrates.config.impl.locale.ErrorKeys;
+import com.badbones69.crazycrates.config.impl.locale.MiscKeys;
+import com.badbones69.crazycrates.config.impl.locale.PlayerKeys;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
