+++ src/main/java/com/badbones69/crazycrates/config/impl/locale/MiscKeys.java
@@ -1,4 +1,4 @@
-package com.badbones69.crazycrates.config.impl.messages;
+package com.badbones69.crazycrates.config.impl.locale;
 
 import ch.jalu.configme.Comment;
 import ch.jalu.configme.SettingsHolder;
